+++ src/components/base/form/ErrorMessage.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import { Label } from 'semantic-ui-react';
+
+const ErrorLabel = ({message}) => (
+    <Label basic color='red' pointing >
+        {message} 
+    </Label>
+)
+
+const ErrorMessage = ({ touched, error, warning}) => (
+    <div>
+        {touched &&
+            ((error && <ErrorLabel message={error}/>) ||
+            (warning && <ErrorLabel message={warning}/>))}
+    </div>
+)
+
+
+export default ErrorMessage;
