+++ src/components/base/form/index.js
@@ -1,0 +1,7 @@
+import ControlInput from './ControlInput';
+import ControlCheckbox from './ControlCheckbox';
+
+export {
+    ControlInput,
+    ControlCheckbox
+};
